add test name to sauce name
